requirements.txt
cachetools==6.2.3
cloudinary==1.44.1
deprecated==1.3.1
email-validator==2.3.0
flask==3.1.2
flask-caching==2.3.1
flask-cors==6.0.2
flask-jwt-extended==4.7.1
flask-limiter==4.1.1
flask-mail==0.10.0
flask-migrate==4.1.0
flask-pydantic-spec==0.8.7
flask-sqlalchemy==3.1.1
flask-wtf==1.2.2
gunicorn==23.0.0
phonenumbers==9.0.20
psycopg2-binary==2.9.11
pycountry==24.6.1
python-dotenv==1.2.1
python-slugify==8.0.4
quas-docs==0.1.2
quas-utils==0.0.6
requests==2.32.5
rich==14.2.0
typing-extensions==4.15.0